test(scheduler): split targeted cancellation coverage
